Rename loss_mitigation.py to photon_recycling.py
