+++ perceval/runtime/error_mitigation/photon_recycling.py
@@ -39,7 +39,7 @@ from perceval.utils.logging import get_logger, channel
 
 from ..computation import Computation
 from .abstract_mitigation import AbstractMitigation
-from ._loss_mitigation_utils import _gen_lossy_dists, _get_avg_exp_from_uni_dist, _generate_one_photon_per_mode_mapping
+from .utils._photon_recycling import _gen_lossy_dists, _get_avg_exp_from_uni_dist, _generate_one_photon_per_mode_mapping
 
 
 def _validate_noisy_input(noisy_input: BSCount | BSDistribution, ideal_photon_count: int):
